backend/pyproject.toml
[project]
name = "test_e2e_phase_9"
version = "0.1.0"
description = "Django REST Framework API"
requires-python = ">=3.12"
dependencies = [
    "django==5.2.7",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "django-environ>=0.11.0",
    "django-cors-headers>=4.3.0",
    "psycopg2-binary>=2.9.0",
    "gunicorn>=21.2.0",
    "whitenoise>=6.6.0",
    "drf-spectacular>=0.27.0",
"redis>=5.0.0",
    "django-redis>=5.4.0",
"celery>=5.3.0",
    "celery[redis]>=5.3.0",
"channels>=4.0.0",
    "channels-redis>=4.1.0",
    "daphne>=4.0.0",
"boto3>=1.34.0",
    "django-storages>=1.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-django>=4.7.0",
    "pytest-cov>=4.1.0",
    "factory-boy>=3.3.0",
    "faker>=22.0.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "django-stubs>=4.2.0",
    "djangorestframework-stubs>=3.14.0",
    "django-debug-toolbar>=4.2.0",
    "ipython>=8.18.0",
]

[tool.black]
line-length = 100
target-version = ['py312']
exclude = '''
/(
    \.git
  | \.venv
  | migrations
  | __pycache__
)/
'''

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "migrations",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # function calls in argument defaults
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--cov=. --cov-report=html --cov-report=term-missing"

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
strict = false
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings"
